Script for creating a spack-based config.sh
Inspired by Nate's do-config.sh, this script creates a
config.sh for configuring camellia2 by hand.  The sript assumes
that spack has already installed camellia2 and then creates a
configure script to be equivalent to spack's installed version.
This method is more useful for development as one can tweak
the configuration automatically and place it in a more
convenient (non-hashed) location.