Refinements to workflow discussion
Spack's strength in creating and managing unique builds with hashes
is fairly incompatible with python's desire to have everything into
a single path.  To date, it seems that the best way to handle this is to
use the module system.  This modifies the configuration scripts and
discussion to match this workflow.  bash and tcsh gave me different pain
points in developing these config files.  Urgh.